pyproject.toml
# SPDX-FileCopyrightText: 2023 osfanbuff63

# SPDX-License-Identifier: CC0-1.0

[tool.poetry]
name = "pkw-tracking-bot"
version = "0.1.0"
description = "A Discord bot to track Parkour Warrior Dojo times from MCC Island."
authors = ["osfanbuff63 <osfanbuff63@osfanbuff63.tech>"]
license = "Apache-2.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
discord-py = "^2.3.2"
arrow = "^1.3.0"
tomlkit = "^0.12.3"
flask = "^3.0.0"

[tool.poetry.group.dev.dependencies]
black = "^23.11.0"
isort = "^5.13.1"
mypy = "^1.7.1"

[tool.poetry.scripts]
pkw-tracking-bot = "pkw_tracking_bot"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
